requirements.txt
torch>=1.9.0
transformers>=4.5.0
datasets>=1.8.0
pandas>=1.3.0
numpy>=1.19.5
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
nltk>=3.6.2
pyyaml>=5.4.1
evaluate>=0.3.0
